+++ kernel/modules/core/SovereignPythonVM.c
@@ -8,7 +8,8 @@
  * =========================================================================
  */
 
-#include "../../../include/sigma_kernel.h"
+#include "../../../include/sigma_libc.h"
+#include "../../../include/SigmaOOP.h"
 #include "../../../include/SovereignPythonZenith.h"
 
 // -------------------------------------------------------------------------